requirements.txt
asgiref==3.5.0
beautifulsoup4==4.11.1
bottle==0.12.19
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
commandlines==0.4.1
Django==4.0.4
djangorestframework==3.13.1
filelock==3.6.0
filetype==1.0.10
Flask==2.1.1
flatbuffers==2.0
gdown==4.4.0
hsh==1.1.0
idna==3.3
imageio==2.16.2
itsdangerous==2.1.2
Jinja2==3.1.1
llvmlite==0.38.0
MarkupSafe==2.1.1
networkx==2.8
numba==0.55.1
numpy==1.22.3
ort-nightly==1.12.0.dev20220412001
packaging==21.3
Pillow==9.1.0
protobuf==3.20.0
PyMatting==1.1.5
pyparsing==3.0.8
PySocks==1.7.1
pyTelegramBotAPI==4.5.0
pytz==2022.1
PyWavelets==1.3.0
rembg==1.0.27
requests==2.27.1
scikit-image==0.19.2
scipy==1.8.0
six==1.16.0
soupsieve==2.3.2
sqlparse==0.4.2
tifffile==2022.4.8
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
typing_extensions==4.1.1
tzdata==2022.1
urllib3==1.26.9
waitress==2.1.1
Werkzeug==2.1.1
